Add files via upload
This is the original design and model for the software aspect of the robot. All of the robot components will be connected into one system that will allow not only the gathering of information as the robot sweeps the pipes, but it will also allow for dynamic inspection and monitoring of the robot. The Software would allow for human intervention and allow for the precise targeting of leakages.